ExponentialSearch.java
Exponential Search involves two steps, Finding a range where the element may be present & To do a binary search on the found range. The idea is to start with subarray size 1 compare its last element with x, then try size 2, then 4 and so on until last element of a sub array is not greater. Once we find an index i (after repeated doubling of i), we know that the element must be present between i/2 and i (Why i/2? because we could not find a greater value in previous iteration)

Time Complexity is 0(Logn)

Applications; used during unbound searches, where size of array is infinite. It works better than Binary Search, where the to be searched element is closer to the first element.