Union Operation with Hashing
Union (list1, list2)
Initialize the result list as NULL and create an empty hash table. Traverse both lists one by one, for each element being visited, look the element in the hash table. If the element is not present, then insert the element to the result list. If the element is present, then ignore it.