Update ArrayLeftRotation.java
This implementation uses the System.arraycopy method to copy the elements of the arr array to the dst array. The first call to System.arraycopy copies the elements from index n to the end of the arr array to the beginning of the dst array. The second call copies the elements from the beginning of the arr array to index n - 1 to the end of the dst array.